Move small components to separate files
