+++ src/header/Title.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import Text from '../text/Text';
+
+const Title = (props) => {
+  const {
+    text,
+  ...attributes,
+  } = props;
+
+  return (
+    <Text
+      numberOfLines={1}
+      {...attributes}
+    >
+      {text}
+    </Text>
+  );
+};
+
+export default Title;
